examples: split AuthWithSignedCmsBase64Example into homo/ and prod/ subpackages
Motivation:
- Provide clear separation of examples using signed CMS Base64 for both
  homologation and production environments.
- Eliminate commented-out lines and manual switching that could confuse
  developers.

Changes:
- Added new 'examples.homo.AuthWithSignedCmsBase64Example' configured
  with ApiEnvironment.HOMO and WSAA homologation endpoint.
- Moved original example into 'examples.prod.AuthWithSignedCmsBase64Example'
  configured with ApiEnvironment.PROD and WSAA production endpoint.
- Documented both homologation and production WSAA endpoints as comments
  for reference.

Impact:
- Developers can now directly run either homo or prod version without
  editing code.
- Improves clarity and plug-and-play usage for Base64 CMS authentication.
